Refactor a Java package from a directory.
The script will recursively walk a directory and replace all instances of a package in all files.  It should be noted
that this does not yet behave well with binary files.

The final step is to traverse the directory tree and rename directories to the new package structure.

An example of usage is as follows:

    # python3 repackage.py -d /tmp --old com.example.old --new com.example.new